util/path.py: break cycle with spack.config
The filter_padding context manager was the only thing in spack.util.path
that needed spack.config: it reads config:install_tree:padded_length.
Padding is an install-tree concept, so move it to spack.store, which
already imports config and util.path. The pure padding_filter callables
stay in util/path.
